Pass audit even if some modules require updates
The audit is intended to notify developers about any modules with
security vulnerabilities. Most can be updated, but some cannot. When
there are modules that cannot be updated, there's really not much that
can be done about it, so we don't want to fail the whole "npm test"
command as a result.